"Add renderTodos function to generate todo items and update Main component"
This diff adds a new function called `renderTodos` to the `Main` component which generates an array of 20 todo items. The implementation of the `Main` component is also updated to use this new function and display the todo items. Additionally, the diff includes some changes to the layout and styling of the todo items.